+++ Tests/StreamChatUITests/SnapshotTests/ChatChannelList/ChatChannelListItemView_Tests.swift
@@ -370,6 +370,7 @@ final class ChatChannelListItemView_Tests: XCTestCase {
 
         view.addSizeConstraints()
         view.components = .mock
+        view.appearance.formatters.channelListMessageTimestamp = DefaultMessageTimestampFormatter()
 
         view.content = .init(
             channel: channel(
@@ -770,7 +771,7 @@ final class ChatChannelListItemView_Tests: XCTestCase {
 
     // MARK: - Timestamp
 
-    func test_timestampText_isNil_whenPreviewMessageIsNil() {
+    func test_timestampText_whenPreviewMessageIsNil_thenTimestampIsNil() {
         let channel: ChatChannel = .mock(
             cid: .unique,
             previewMessage: nil
@@ -781,20 +782,17 @@ final class ChatChannelListItemView_Tests: XCTestCase {
         XCTAssertNil(itemView.timestampText)
     }
 
-    func test_timestampText_whenPreviewMessageExists() {
+    func test_timestampText_whenPreviewMessageExists_thenUsesCreatedAtFromPreviewMessage() {
         let channel: ChatChannel = .mock(
             cid: .unique,
             previewMessage: .mock(
-                id: .unique,
-                cid: .unique,
-                text: .unique,
-                author: .mock(id: .unique),
                 createdAt: Date(timeIntervalSince1970: 1)
             )
         )
 
         let itemView = ChatChannelListItemView()
         itemView.content = .init(channel: channel, currentUserId: nil)
+        itemView.appearance.formatters.channelListMessageTimestamp = DefaultMessageTimestampFormatter()
 
         XCTAssertEqual(
             itemView.timestampText,
@@ -802,7 +800,7 @@ final class ChatChannelListItemView_Tests: XCTestCase {
         )
     }
 
-    func test_timestampText_whenSearchingMessage() {
+    func test_timestampText_whenSearchingMessage_thenUsesCreatedAtFromSearchResultMessage() {
         let itemView = ChatChannelListItemView()
         itemView.content = .init(
             channel: .mockNonDMChannel(previewMessage: nil),
@@ -812,10 +810,91 @@ final class ChatChannelListItemView_Tests: XCTestCase {
                 message: .mock(text: "Some text", createdAt: Date(timeIntervalSince1970: 1))
             )
         )
+        itemView.appearance.formatters.channelListMessageTimestamp = DefaultMessageTimestampFormatter()
 
         XCTAssertEqual(
             itemView.timestampText,
             "12:00 AM"
+        )
+    }
+
+    func test_timestampText_whenCreatedAtIsToday_thenShowsTimeOnly() {
+        let channel: ChatChannel = .mock(
+            cid: .unique,
+            previewMessage: .mock(
+                createdAt: Date(timeIntervalSince1970: 1)
+            )
+        )
+
+        let mockCalendar = Calendar_Mock()
+        mockCalendar.mockIsDateInToday = true
+        let formatter = ChannelListMessageTimestampFormatter()
+        formatter.calendar = mockCalendar
+
+        let itemView = ChatChannelListItemView()
+        itemView.content = .init(channel: channel, currentUserId: nil)
+        itemView.appearance.formatters.channelListMessageTimestamp = formatter
+
+        XCTAssertEqual(
+            itemView.timestampText,
+            "12:00 AM"
+        )
+    }
+
+    func test_timestampText_whenCreatedAtIsYesterday_thenShowsYesterday() {
+        let channel: ChatChannel = .mock(
+            cid: .unique,
+            previewMessage: .mock(
+                createdAt: Calendar.current.date(byAdding: .day, value: -1, to: Date()) ?? Date()
+            )
+        )
+
+        let itemView = ChatChannelListItemView()
+        itemView.content = .init(channel: channel, currentUserId: nil)
+
+        XCTAssertEqual(
+            itemView.timestampText,
+            "Yesterday"
+        )
+    }
+
+    func test_timestampText_whenCreatedAtInLastWeek_thenShowsWeekDay() {
+        let channel: ChatChannel = .mock(
+            cid: .unique,
+            previewMessage: .mock(
+                createdAt: Date(timeIntervalSince1970: 1_690_998_292)
+            )
+        )
+
+        let mockCalendar = Calendar_Mock()
+        mockCalendar.mockIsDateInLastWeek = true
+        let formatter = ChannelListMessageTimestampFormatter()
+        formatter.calendar = mockCalendar
+
+        let itemView = ChatChannelListItemView()
+        itemView.content = .init(channel: channel, currentUserId: nil)
+        itemView.appearance.formatters.channelListMessageTimestamp = formatter
+
+        XCTAssertEqual(
+            itemView.timestampText,
+            "Wednesday"
+        )
+    }
+
+    func test_timestampText_whenCreatedAtBeforeLastWeek_thenShowsDate() {
+        let channel: ChatChannel = .mock(
+            cid: .unique,
+            previewMessage: .mock(
+                createdAt: Date(timeIntervalSince1970: 1_690_998_292)
+            )
+        )
+
+        let itemView = ChatChannelListItemView()
+        itemView.content = .init(channel: channel, currentUserId: nil)
+
+        XCTAssertEqual(
+            itemView.timestampText,
+            "8/2/23"
         )
     }
 
@@ -1034,6 +1113,7 @@ final class ChatChannelListItemView_Tests: XCTestCase {
         let view = ChatChannelListItemView().withoutAutoresizingMaskConstraints
         view.components = components
         view.appearance = appearance
+        view.appearance.formatters.channelListMessageTimestamp = DefaultMessageTimestampFormatter()
         view.content = content
         view.addSizeConstraints()
         return view
@@ -1061,5 +1141,23 @@ private extension ChatChannelListItemView {
         NSLayoutConstraint.activate([
             widthAnchor.constraint(equalToConstant: 400)
         ])
+    }
+}
+
+private class Calendar_Mock: ChannelListMessageTimestampCalendar {
+    var mockIsDateInToday = false
+    var mockIsDateInYesterday = false
+    var mockIsDateInLastWeek = false
+
+    func isDateInToday(_ date: Date) -> Bool {
+        mockIsDateInToday
+    }
+
+    func isDateInYesterday(_ date: Date) -> Bool {
+        mockIsDateInYesterday
+    }
+
+    func isDateInLastWeek(_ date: Date) -> Bool {
+        mockIsDateInLastWeek
     }
 }